[ADD] 카테고회 조회 코드 추가 및 테스트 코드 추가

+++ src/main/java/com/f3f/community/category/service/CategoryService.java
@@ -7,7 +7,6 @@ import com.f3f.community.exception.categoryException.*;
 import com.f3f.community.post.domain.Post;
 import com.f3f.community.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,6 +14,8 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import static com.f3f.community.common.constants.ResponseConstants.OK;
 
 @Service
 @RequiredArgsConstructor
@@ -26,22 +27,10 @@ public class CategoryService {
 
     @Transactional
     public Long createCategory(SaveRequest saveRequest) throws Exception {
-        if (saveRequest.getCategoryName() == null) {
-            throw new NotFoundCategoryNameException();
-        }
-        if (saveRequest.getCategoryName().equals("")) {
-            throw new NotFoundCategoryNameException("이름이 빈 스트링입니다.");
-        }
-        if (saveRequest.getPostList() == null) {
-            throw new NotFoundCategoryPostListException();
-        }
-        if (saveRequest.getChildCategory() == null) {
-            throw new NotFoundChildCategoryListException();
-        }
         if (categoryRepository.existsByCategoryName(saveRequest.getCategoryName())) {
             throw new DuplicateCategoryNameException();
         }
-        if (saveRequest.getParents() == null && categoryRepository.existsByCategoryName("root")) {
+        if (saveRequest.getParents() == null && !saveRequest.getCategoryName().equals("root")) {
             throw new NotFoundParentCategoryException();
         }
         Category category = saveRequest.toEntity();
@@ -59,6 +48,36 @@ public class CategoryService {
         return category.getId();
     }
 
+    @Transactional(readOnly = true)
+    public List<Category> getChildCategories(Long catId) {
+        Category category = categoryRepository.findById(catId).orElseThrow(NotFoundCategoryByIdException::new);
+        return categoryRepository.findCategoriesByParents(category);
+    }
+
+    @Transactional(readOnly = true)
+    public List<Post> getPostsOfChild(Long catId) {
+        List<Post> result = new ArrayList<>();
+        for (Category cat : getChildCategories(catId)) {
+            result.addAll(cat.getPostList());
+        }
+        return result;
+    }
+
+    @Transactional(readOnly = true)
+    public List<Category> getChildsOfRoot() {
+        Category root = categoryRepository.findByCategoryName("root").orElseThrow(NotFoundCategoryByNameException::new);
+        return getChildCategories(root.getId());
+    }
+
+    @Transactional(readOnly = true)
+    public List<Post> getPostsOfRootChild() {
+        List<Post> result = new ArrayList<>();
+        for (Category cat : getChildsOfRoot()) {
+            result.addAll(cat.getPostList());
+        }
+        return result;
+    }
+
     // 기존에 짯던 포스트리스트만 리턴하는 getPost입니다
 //    @Transactional(readOnly = true)
 //    public List<Post> getPosts(Long catId) {
@@ -72,26 +91,26 @@ public class CategoryService {
 //    }
 
     // 리팩터링해서 깊이를 키로 가지고, 해당 깊이에 포스트리스트를 밸류로 가지는 해쉬맵을 리턴하게 변경하였습니다
-    @Transactional(readOnly = true)
-    public Map<Long,List<Post>> getPosts(Long catId) {
-        Category category = categoryRepository.findById(catId).orElseThrow(NotFoundCategoryByIdException::new);
-        List<Post> posts = new ArrayList<>(category.getPostList());
-        Map<Long, List<Post>> result = new HashMap<>();
-        result.put(category.getDepth(), posts);
-
-
-        for (Category child : category.getChildCategory()) {
-            Map<Long, List<Post>> temp = getPosts(child.getId());
-            for (Long key : temp.keySet()) {
-                if (!result.containsKey(key)) {
-                    result.put(key, new ArrayList<>());
-                }
-                result.get(key).addAll(temp.get(key));
-            }
-        }
-
-        return result;
-    }
+//    @Transactional(readOnly = true)
+//    public Map<Long, List<Post>> getPosts(Long catId) {
+//        Category category = categoryRepository.findById(catId).orElseThrow(NotFoundCategoryByIdException::new);
+//        List<Post> posts = new ArrayList<>(category.getPostList());
+//        Map<Long, List<Post>> result = new HashMap<>();
+//        result.put(category.getDepth(), posts);
+//
+//
+//        for (Category child : category.getChildCategory()) {
+//            Map<Long, List<Post>> temp = getPosts(child.getId());
+//            for (Long key : temp.keySet()) {
+//                if (!result.containsKey(key)) {
+//                    result.put(key, new ArrayList<>());
+//                }
+//                result.get(key).addAll(temp.get(key));
+//            }
+//        }
+//
+//        return result;
+//    }
 
     @Transactional
     public String updateCategoryName(Long catId, String newName) {
@@ -100,15 +119,19 @@ public class CategoryService {
             throw new DuplicateCategoryNameException();
         } else {
             category.updateName(newName);
-            return "ok";
+            return OK;
         }
     }
 
     @Transactional
     public String deleteCategory(Long catId) {
         Category category = categoryRepository.findById(catId).orElseThrow(NotFoundCategoryByIdException::new);
-        categoryRepository.delete(category);
-        return "ok";
+        if (category.getChildCategory().isEmpty()) {
+            categoryRepository.delete(category);
+        }else {
+            throw new NotEmptyChildCategoryException();
+        }
+        return OK;
     }
 
 }
